backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-image==0.22.0
Pillow==10.3.0
pytesseract==0.3.10
pymupdf==1.24.1
reportlab==4.2.0
torch==2.2.2
torchvision==0.17.2
httpx==0.27.0
aiofiles==23.2.1
python-jose==3.3.0
